feat:support service alias from xds (#1070)
