+++ apiserver/httpserver/docs/circuitbreaker.go
@@ -838,60 +838,4 @@ PUT /naming/v1/circuitbreaker/rules/enable
 }
 ~~~
 `
-
-	enrichCreateRouterRuleApiNotes = `
-创建路由规则
-
-~~~
-POST /naming/v2/routings
-
-# 开启北极星服务端针对控制台接口鉴权开关后，需要添加下面的 header
-Header X-Polaris-Token: {访问凭据}
-
-~~~
-`
-	enrichDeleteRouterRuleApiNotes = `
-删除路由规则
-
-~~~
-DELETE /naming/v2/routings
-
-# 开启北极星服务端针对控制台接口鉴权开关后，需要添加下面的 header
-Header X-Polaris-Token: {访问凭据}
-
-~~~
-`
-	enrichUpdateRouterRuleApiNotes = `
-更新路由规则
-
-~~~
-PUT /naming/v2/routings
-
-# 开启北极星服务端针对控制台接口鉴权开关后，需要添加下面的 header
-Header X-Polaris-Token: {访问凭据}
-
-~~~
-`
-	enrichGetRouterRuleApiNotes = `
-获取路由规则
-
-~~~
-GET /naming/v2/routings
-
-# 开启北极星服务端针对控制台接口鉴权开关后，需要添加下面的 header
-Header X-Polaris-Token: {访问凭据}
-
-~~~
-`
-	enrichEnableRouterRuleApiNotes = `
-更新路由规则
-
-~~~
-PUT /naming/v2/routings
-
-# 开启北极星服务端针对控制台接口鉴权开关后，需要添加下面的 header
-Header X-Polaris-Token: {访问凭据}
-
-~~~
-`
 )
