+++ apiserver/httpserver/docs/naming_client_access_apidoc.go
@@ -21,6 +21,7 @@ import (
 	"github.com/emicklei/go-restful/v3"
 	restfulspec "github.com/polarismesh/go-restful-openapi/v2"
 	apiservice "github.com/polarismesh/specification/source/go/api/v1/service_manage"
+	apitraffic "github.com/polarismesh/specification/source/go/api/v1/traffic_manage"
 )
 
 var (
@@ -57,4 +58,41 @@ func EnrichDiscoverApiDocs(r *restful.RouteBuilder) *restful.RouteBuilder {
 	return r.Doc("服务发现").
 		Metadata(restfulspec.KeyOpenAPITags, registerInstanceApiTags).
 		Notes(enrichDiscoverApiNotes)
+}
+
+func enrichCreateRoutingsApiDocs(r *restful.RouteBuilder) *restful.RouteBuilder {
+	return r.Doc("创建路由规则(V2)").
+		Metadata(restfulspec.KeyOpenAPITags, routingRulesApiTags).
+		Operation("v2CreateRoutings").
+		Reads([]apitraffic.RouteRule{}).
+		Notes(enrichCreateRouterRuleApiNotes)
+}
+
+func enrichDeleteRoutingsApiDocs(r *restful.RouteBuilder) *restful.RouteBuilder {
+	return r.Doc("删除路由规则(V2)").
+		Metadata(restfulspec.KeyOpenAPITags, routingRulesApiTags).
+		Operation("v2DeleteRoutings").
+		Notes(enrichDeleteRouterRuleApiNotes)
+}
+
+func enrichUpdateRoutingsApiDocs(r *restful.RouteBuilder) *restful.RouteBuilder {
+	return r.Doc("更新路由规则(V2)").
+		Metadata(restfulspec.KeyOpenAPITags, routingRulesApiTags).
+		Operation("v2UpdateRoutings").
+		Reads([]apitraffic.RouteRule{}).
+		Notes(enrichUpdateRouterRuleApiNotes)
+}
+
+func enrichGetRoutingsApiDocs(r *restful.RouteBuilder) *restful.RouteBuilder {
+	return r.Doc("获取路由规则(V2)").
+		Metadata(restfulspec.KeyOpenAPITags, routingRulesApiTags).
+		Operation("v2GetRoutings").
+		Notes(enrichGetRouterRuleApiNotes)
+}
+
+func enrichEnableRoutingsApiDocs(r *restful.RouteBuilder) *restful.RouteBuilder {
+	return r.Doc("启用路由规则(V2)").
+		Metadata(restfulspec.KeyOpenAPITags, routingRulesApiTags).
+		Operation("v2EnableRoutings").
+		Notes(enrichEnableRouterRuleApiNotes)
 }
