+++ apiserver/xdsserverv3/server_test.go
@@ -278,7 +278,7 @@ func TestNodeHashID(t *testing.T) {
 					},
 				},
 			},
-			TargetID: "sidecar~default/" + TLSModeStrict,
+			TargetID: "default/" + TLSModeStrict,
 		},
 		{
 			Node: &core.Node{
@@ -293,7 +293,7 @@ func TestNodeHashID(t *testing.T) {
 					},
 				},
 			},
-			TargetID: "sidecar~polaris/" + TLSModePermissive,
+			TargetID: "polaris/" + TLSModePermissive,
 		},
 		{
 			Node: &core.Node{
@@ -308,7 +308,7 @@ func TestNodeHashID(t *testing.T) {
 					},
 				},
 			},
-			TargetID: "sidecar~default",
+			TargetID: "default",
 		},
 		// bad case: wrong tls mode
 		{
@@ -324,14 +324,14 @@ func TestNodeHashID(t *testing.T) {
 					},
 				},
 			},
-			TargetID: "sidecar~default",
+			TargetID: "default",
 		},
 		// no node metadata
 		{
 			Node: &core.Node{
 				Id: "default/9b9f5630-81a1-47cd-a558-036eb616dc71~172.17.1.1",
 			},
-			TargetID: "sidecar~default",
+			TargetID: "default",
 		},
 		// metadata does not contain tls mode kv
 		{
@@ -347,7 +347,7 @@ func TestNodeHashID(t *testing.T) {
 					},
 				},
 			},
-			TargetID: "sidecar~default",
+			TargetID: "default",
 		},
 		{
 			Node: &core.Node{
@@ -449,7 +449,7 @@ func TestSnapshot(t *testing.T) {
 	resp := x.namingServer.CreateRoutingConfigsV2(discoverSuit.DefaultCtx, mockRouterRule)
 	assert.Equal(t, apimodel.Code_ExecuteSuccess, apimodel.Code(resp.GetCode().GetValue()))
 	x.namingServer.Cache().TestUpdate()
-	x.pushRegistryInfoToXDSCache(sis)
+	x.pushSidecarInfoToXDSCache(sis)
 	x.pushGatewayInfoToXDSCache(sis)
 
 	snapshot, _ := x.cache.GetSnapshot("default")
